requirements.txt
google-genai
python-docx
rich
pydantic
PyYAML
pytest
black
ruff
